QRTool/VirusTotalAPI/requirements.txt
requests
load_dotenv